pixelart_tool/requirements.txt
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.7.0
matplotlib>=3.7.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.8
scikit-learn>=1.3.0